+++ src/lib/components/variant/A11yHorizontalWrapper.svelte
@@ -25,6 +25,7 @@
 	@media (min-width: 768px) {
 		.isOpen {
 			display: grid;
+			gap: 5px;
 			grid-template:
 				'picker . . a11y'
 				'input input input a11y';
@@ -42,10 +43,6 @@
 	div :global(.a11y-notice) {
 		grid-area: a11y;
 		margin: 0 4px 0 6px;
-	}
-
-	div :global(.slider-wrapper) {
-		margin-left: 5px;
 	}
 
 	.isPopup {